refactor: move patch optimizer code into separate file
