+++ packages/sync/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./encoding";
 export * from "./patch-state-machine";
+export * from "./patch-optimizer";
 export * from "./patch";
