+++ packages/sync/src/patch-state-machine.ts
@@ -1,6 +1,8 @@
 import type { ReadonlyDeep } from "type-fest";
 import type { Patch, PatchPath } from "./patch";
 import { dedupePatch } from "./patch-deduper";
+import type { EntityPatchOptimizerRecord } from "./patch-optimizer";
+import { optimizePatchOperationValue } from "./patch-optimizer";
 
 /**
  * A state machine that records all state changes made as atomic patches,
@@ -189,54 +191,6 @@ function createEntityRepository<
 
   return entity;
 }
-
-export type EntityPatchOptimizerRecord<State extends PatchableState> = {
-  [EntityName in keyof State]?: EntityPatchOptimizer<
-    State[EntityName][keyof State[EntityName]]
-  >;
-};
-
-export type EntityPatchOptimizer<Entity> = {
-  [K in keyof Entity]?: PropertyPatchOptimizer<Entity, K>;
-};
-
-export interface PropertyPatchOptimizer<
-  Entity,
-  Key extends keyof Entity = keyof Entity,
-> {
-  filter?: PropertyPatchOptimizerFilter<Entity[Key]>;
-  transform?: (value: Entity[Key]) => Entity[Key];
-}
-
-export type PropertyPatchOptimizerFilter<Value> = (
-  newValue: Value,
-  oldValue: Value,
-) => boolean;
-
-function optimizePatchOperationValue<Entity, Key extends keyof Entity>(
-  {
-    transform,
-    filter = defaultOptimizerFilter,
-  }: PropertyPatchOptimizer<Entity, Key> | undefined = empty,
-  newValue: Entity[Key],
-  oldValue: Entity[Key],
-): { value: Entity[Key] } | undefined {
-  if (transform) {
-    newValue = transform(newValue);
-    oldValue = transform(oldValue);
-  }
-  if (!filter(newValue, oldValue)) {
-    return;
-  }
-  return { value: newValue };
-}
-
-const empty = Object.freeze({});
-
-const defaultOptimizerFilter: PropertyPatchOptimizerFilter<unknown> = (
-  newValue,
-  oldValue,
-) => newValue !== oldValue;
 
 function createFullStatePatch<State extends PatchableState>(
   state: State,
